+++ dsm2/src/hdf_tidefile/dsm2_tidefile_input_storage_virtual_xsect.cpp
@@ -353,7 +353,7 @@ void virtual_xsect_number_rows_hdf5_f(const hid_t *file_id, hsize_t* nrecords, i
 
     
 /** get one row worth of information from the buffer */
-void virtual_xsect_query_from_buffer_f(size_t* row, 
+void virtual_xsect_query_from_buffer_f(int32_t* row, 
                         int * a_chan_no,int * a_num_virt_sec,int * a_vsecno,int * a_num_elev,double * a_min_elev,double * a_elevation,double * a_area,double * a_wet_p,double * a_width, int * ierror
                         )
 {
